sis/requirements.txt
Pillow
h5py
tensorflow
Keras
Flask
opencv-python
imagehash
scikit-image
scikit-learn
imutils

